requirements.txt
fastapi
uvicorn
websockets
sqlalchemy
pymysql
redis
opencv-python
ultralytics
aiortc
av
pydantic
pydantic-settings
pytest
requests
allure-pytest
psutil
allure-python-commons
websocket-client
